rename radioOption -> radio-option
